Add a diagnosis hint to pause point wait timeouts
Agents found timeout root-causing hard: the same error covers a missed code
path, a stopped PlayMode, and an already-paused Editor. The timeout error now
carries details.hint with a deterministic diagnosis for those three states so
the next action is obvious without extra status probing.
